renamed StableMotifs to PyStableMotifs to avoid package conflict
