+++ Examples and Tutorials/ControlExample1.py
@@ -1,5 +1,5 @@
 import PyBoolNet
-import StableMotifs as sm
+import PyStableMotifs as sm
 import networkx as nx
 from timeit import default_timer
 print("Loading network . . .")
