+++ PyStableMotifs/DomainOfInfluence.py
