+++ PyStableMotifs/Format.py
