+++ PyStableMotifs/TimeReversal.py
