--- StableMotifs/__init__.py
@@ -1,12 +1,0 @@
-import PyBoolNet
-
-from StableMotifs import Format
-from StableMotifs import DomainOfInfluence
-from StableMotifs import Reduction
-from StableMotifs import TimeReversal
-from StableMotifs import RestrictSpace
-from StableMotifs import Succession
-from StableMotifs import RandomBooleanNetworks
-from StableMotifs import Attractor
-from StableMotifs.AttractorRepertoire import AttractorRepertoire
-from StableMotifs import Export
